Revert "Improve single-element detection (#4)"
This reverts commit 09766e3cec845dadc5dc185eccdd3ec2caaf91f6.

Any other modern feature used in this fork can either be transpiled by
Babel or polyfilled for IE 11. But nodes being an instance of the
EventTarget constructor cannot be polyfilled. And IE 11 does not expose
EventTarget as a constructor.
Reverting the detection to the method of the EventTarget interface
avoids relying on a non-polyfillable API.