Improve type inference for the listener callback
Instead of using a generic Event type for the argument, the map of event types to the corresponding Event object types is used by the inference.

This is a BC break for projects specifying the generic types explicitly, as it changes the generic signature. For project relying on type inference, the inferred type becomes more specific.