bugfix: Allow delegation to any Node
For use in custom elements with a ShadowRoot, it may be desirable to delegate
event handling to the ShadowRoot, as that is the root Node for all elements
contained within it.

ShadowRoot is a descendent of Node, and the .contains()-method is
defined on Node, such that checking that the currentTarget where the
delegated event handler is active is any Node rather than just an
Element enables this.